Move classes and interfaces from SettingsService.cs into own files (one file per class/interface)
