+++ Template10 (Library)/Services/SettingsService/IPropertyMapping.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Template10.Services.SettingsService
+{
+    public interface IPropertyMapping
+    {
+        IStoreConverter GetConverter(Type type);
+    }
+}
