+++ Template10 (Library)/Services/SettingsService/IStoreConverter.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Template10.Services.SettingsService
+{
+    public interface IStoreConverter
+    {
+        string ToStore(object value, Type type);
+        object FromStore(string value, Type type);
+    }
+}
