+++ Template10 (Library)/Services/SettingsService/JsonMapping.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Template10.Services.SettingsService
+{
+    public class JsonMapping : IPropertyMapping
+    {
+        protected IStoreConverter jsonConverter = new JsonConverter();
+        public IStoreConverter GetConverter(Type type) => this.jsonConverter;
+    }
+}
